+++ packages/web/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.0.3 - 2020-07-27
 
 - The `createApp` router argument is now typed as `Router`.
+- `GetUsersByIdParams` has been temporary typed as `any` to prevent `ESLint` error.
 
 ## 2.0.2 - 2020-06-30